BUG: Fix a template name formation in python wrappings
FIXES: ITK-3378

This patch fixes the formation of names for certain template classes in
Python wrappings, if the namespace differs from "itk" and class name with
a certain template encompasses the name belonging to the namespace.

For example, the formation of names was incorrect for the
following class std::numeric_limits:
     std::numeric_limits, stdnumeric_limitsUC, itsUC

The class name "itsUC" is incorrect. The correct name should be "UC".

This problem more often occurs while using WrapITK as the
external project for wrappings for custom ITK-based classes that might
exist in the namespace different from "itk".

Change-Id: I8f933b6b9f7e9873fe3f5ffda6c28d8fce6643ab