BUG: Preserve backwards compatibility
When comparing itk::Functions, only the operator!=() is
is required to be set.  The use of the AlmostEquals is
not needed here, and forces the need for an operator==()
unnecessarily.

The test case was enhanced to induce a failure for the
obscure case when operator==() is not available. Furthermore,
the SetFunction() member function is explicitly tested.

Change-Id: Ic2fee4ddae72a4215bfaa1e7685789caa2158afa