PERF: Optimize WarpImageFilter by avoiding refcount bouncing
WarpImageFilter::EvaluateDisplacementAtPhysicalPoint() calls
GetDisplacementField() for each output pixel, which uses
itk::ProcessObject::GetInput(). The latter has a conditional
operator with different value types on its two branches, leading
to a temporary SmartPointer object being created then immediately
destroyed.

This could arguably be fixed in GetInput(), however there's no point
in calling that function and doing the range-check in a loop. It's
better to just cache the displacement field pointer.

Our code was using the resampler in the context of the Orfeo ToolBox
and spent 40% of time managing the smart pointer instances (in a
Release build). This optimization reduced the run time by 70%.

Change-Id: I705d6f66a1c761422a23b8ada84ef090912c1467