ENH: Allow projects to build against an external module.
Since ITK modules are dependent on each other, the build artifacts, e.g. the
library binaries, from the externally built module are output to the ITK build
tree.

To build against an externally built module, just build against the
ITK configuration, e.g.

  find_package(ITK REQUIRED COMPONENTS AnExternallyBuiltModule)

A new entry is added to the ModuleInfo with an optional TARGETS_FILE for the
module. This file will be included by ITKConfig.cmake to tell the build system
the location of imported libraries.

Change-Id: Idac4c5ceb054548d2bb67a2900b06a890d301bfa