ENH: Better layout strategy
Orignally QuickView only allowed a single row of viewports. This patch
enhances the viewport layout algorithm so that the number of columns
can be specified as well as the size of each viewport. The enhancements
are backward compatible.

A test is included that exercises the new SetNumberOfColumns and
SetViewPortSize methods.

Change-Id: Ib7828d29167aa57d2a2838aa7d4fb61839d81728