BUG: Demonstrate side effect on registration due to BSplineScatter change
This patch shows a regression in convergence monitoring from changes
in the BSplineScatteredDataPointSetToImageFilter in change:
51c2ff58e04a25166b6aafc7d7590c2ae74f2ec6

This regression is detected by the SimpleITK test
sitkRegistrationMethodTest.Optimizer_ScalesEstimator.

Adding a C++ implementation of the SimpleITK test. The input images
are each 2 Gaussian blobs offset by -2.8, 9.5. While the SimpleITK
test does not get too close ( .6 RMS of parameters), with the
regression the optimizer only ran one iteration.

Change-Id: Id88eccf891f76e2b9ce11ce5fc24cc1ca62823f8