BUG: Increase tolerance for PhysicsBasedNonRigidRegistrationMethod
The PhysicsBasedNonRigidRegistrationMethod is failing on the
Debinan-sid-gcc4-32bit dashboard system. There appear to be error
accumulated at the boundary.

Here are the reported image error statistics:
ImageError              1583
ImageError Minimum      3
ImageError Maximum      33
ImageError Mean         4.75995

This patch explicitly set defaults, and increases the Intensity
tolerance from 2 to 4. It's expected that most of the 1583 pixel are
below this value and that a future patch may increase the
NumberOfPixelsTolerance value. This will be preference over increasing
the radius.

Change-Id: I077e22cb557acfce58c76e85d201fcb824d24e01