BUG:  Updating B-spline classes with the same epsilon strategy.
The B-spline parametric domain in 1-D is defined on the half-closed
interval [a,b).  Extension to n-D is defined similarly.  This
presents some difficulty for defining the the image domain to be
co-extensive with the parametric domain.  We use the B-spline
epsilon to push the edge of the image boundary inside the
B-spline parametric domain.

Change-Id: I97260b4a9cde3f4ddd545412bb994605d4a35442