ENH: Optimize AdaptiveHistogramEqualization with MovingHistogram base
Implement the filter as a subclass of the MovingHistogramImageFilter,
providing more efficient updates to a histogram while moving in a
line, and multi-threading.

There can be a significant (70x+) performance improvement with larger
radius and multi-threading. There is reasonable scaling with
multiple threads. However with just one thread, and a 1 radius there
almost a 2x improvement.

The boundary condition has changed. Prior the zero flux boundary
condition was used. Now only the image pixels are used, and the
smaller histogram elements are uniformly weighted.

The use of a temporary float image has been remove in favor of scaling on demand.

The UseLookupTable option has been removed/deprecated.

Change-Id: Ib737905a26578878696a1da9ee10906ce9db3b45