COMP: Use ITK's pygccxml by correctly setting up the sys.path
This was previously not so important but pygccxml's API could change and we
need to make sure the version shipped with ITK is used. For this to work, the
path needs to be correctly inserted in sys.path. It is also cleaner to insert
at position 1, as 0 is the directory containing the script that was used to
invoke the Python interpreter
(https://docs.python.org/2/library/sys.html#sys.path)

Change-Id: I3a0a79ed632b4a0008a7b293d8b9bc40ccb84f8e