BUG: Use random seed with for MersenneTwisterRandomVariateGenerator.
As reported in ITK-3387:

According to the documentation GetInstance() initializes the seed with the wall
clock. But no seed is set, so it is always 121212 as set in the constructor.
ImageRandomConstIteratorWithIndex, and probably at lot of classes, use
GetInstance but do not initialize the seed. This means that an
ImageRandomConstIteratorWithIndex will always walk the image in the same order.
To get random behaviour it is necessary to explicitly call ReinitializeSeed on
the the iterator.

The "obvious" fix is to just let the use call ReinitializeSeed, but this puts
the burden of making the random iterator random on the user, which is error
prone without any benefit. I believe the standard assumption about PRKGs are
that they try to use a random seed, unless explicitly asked not to.

Change-Id: I40dc0c7d4d403653c3b4901517be3f74dda8116c