ENH: Allow operations on Indexes and Offsets in Python
ITK-3326

Using an operation on itk.Index or itk.Offset was failing
because the sequence typemap (defined in pyBase.i) expected
integers. Writing an operation like "index + offset" was
failing with the following error:
  ValueError: Expecting a sequence of int

This is because the sequence is using the long int type
(see the __setitem__ and  __getitem__ methods).

This patch allows to do operations like "index + offset"
directly in Python. The __add__ and all other operation
functions are correctly defined for Index and Offset.

A small test was added to check for different combinations
of operations with indexes and offsets.

Change-Id: I0ae33dc9675a46208826934eadfdb32b34313f0b