BUG: Fix detection of the gold linker.
The detection of the presence of the gold linker was using the
ITK_CHECK_CXX_COMPILER_FLAG macro. However, this macro uses the CMake macro
CHECK_CXX_SOURCE_COMPILES and it was testing the falg with
CMAKE_REQUIRED_DEFINITIONS instead of CMAKE_REQUIRED_FLAGS. But, this is a
linking flag, and the flag is not used in the link line when it is added to
CMAKE_REQUIRED_DEFINITIONS.

Use CMAKE_REQUIRED_FLAGS and CHECK_CXX_SOURCE_COMPILES directly.

Change-Id: I28f94a8452a5cf0a2cd2c59e616128fc61cf0336