ENH: Change BUILD_EXAMPLES default to OFF.
In order to improve the first impressions of new users of the toolkit,
I would like change the default BUILD_EXAMPLES build configuration to
OFF.  This will decrease the default build time, and it the reduced
number of targets will help with Visual Studio builds.

When developing, it is nice to have BUILD_EXAMPLES OFF too, at least
at first, to decrease build time.

Change-Id: I095f5fb072ed846112dd6c25332762c694e6a90e