BUG: Affine transformation was estimated without sufficient number of points.
The Affine transformation was estimated with less than dim+1
points. The code now throws an exception when this is the case. Also
updated the documentation with regard to the results obtained with Versor3D
and Rigid2D when they receive less than dim points. In these cases only the
translational component is estimated. This is strange behavior, but it has been
there for a long time so not going to change the code.

Change-Id: I8034852db4eb83640de21a3e2ccba7c86556c06a