BUG: Valgrind detected an invalid read
The new Math::Compare... methods seem to be accessing too much memory
for some types. In particular RGBPixel<unsigned char>. The tests were
using it.Value() methods to get the values of pixels. This method
returns a reference. Changing the Value() to Get() removes the defect.

Valgrind detected the improper access.

A future patch should address the underlying issues with the compare
methods.

Change-Id: If53a8b830a5f62bf44e0aa80cb8e85812730717c