BUG: Implement SetAutoMinimumMaximum in itkScalarImageToHistogramGenerator
Fixes issue ITK-2958 : when using SetHistogramMin() and SetHistogramMax() the
histogram bin minimum and bin maximum of the internal itkSampleToHistogramFilter
were set but ignored in the case where the AutoMinimumMaximum flag was true.

This commits implements the function SetAutoHistogramMinimumMaximum(), wrapping
a call to itkSampleToHistogramFilter::SetAutoMinimumMaximum(). To manually set
the histogram min and max the user will need to set this flag to false.

Note: a maybe more natural API would be to automatically set the auto flag
to false when setting the min/max range, but this would change the behavior
of itkSampleToHistogramFilter. For now this commit just offers documentation
changes to indicate the required calls.

Change-Id: Id812f2cca03130d057e5f57d8d8403fecce7d2bb