BUG: Ensure Transform IO factories are registered once.
This commit updates the transform factories to ensure they
are registered only once when built as a shared library linked
to both a main application and application plugin dynamically
loaded.

It introduces explicit instantiation where templated factories
are explicitly instantiated for "double" and "float" types.

This will allow templated factories to be registered in one translation
unit and instantiated in an other.

See http://stackoverflow.com/questions/8024010/why-do-template-class-functions-have-to-be-declared-in-the-same-translation-unit

Note that this commit also updates IOTransformDCMTK remote IO
applying similar changes.

Suggested-by: Bradley Lowekamp <blowekamp@mail.nih.gov>

Change-Id: I38e1e6e089cb198032e30d3d3aa64ab426649cf4