BUG: Ensure factory is registered once.
This commit updates CMakeLists.txt to ensure built library
is shared.

This commit introduces explicit instantiation where
templated factories are explicitly instantiated for "double" and "float"
types.

This will allow templated factories to be registered in one translation
unit and instantiated in an other.

See http://stackoverflow.com/questions/8024010/why-do-template-class-functions-have-to-be-declared-in-the-same-translation-unit

See ITK issue #3393
https://issues.itk.org/jira/browse/ITK-3393