ENH:  Accommodate epsilon error on left side.
The B-spline parametric domain in 1-D is defined
on the half-closed interval [a,b).  Extension to n-D is
defined similarly.  This presents some difficulty for
defining the image domain to be co-extensive with the
parametric domain.  We use the B-spline epsilon to push
the edge of the image boundary inside the B-spline parametric
domain.  This change specifically targets the left side
of the domain, i.e. abs( point - a ) < epsilon.

Change-Id: Ife59f8c626624d36324d139cbc70d68e759272d7