BUG: Do not use dynamic_cast for DisplacementFieldType in HDF5 test.
On MacOSX, the dynamic_cast of this templated type, which is not explicitly
instantiated, fails when shared libraries are created with BUILD_SHARED_LIBS.
Use a static_cast instead.

In cases where the type is not know at compile time and needs to be determined
at run time,  ->GetTransformTypeAsString() can be used.

Related to ITK-3393.

Change-Id: Id0f4bf0c04146ecefdf1d2f9884872c981e2f553