ENH: Add Floored and Truncated Modulus to VNL
The % operator and the library function fmod implement
the remainder operation such that the result has the same
sign as the dividend.  While this is a valid mathematical
interpretation, it is also common for the mathematical modulus
to be defined such that the result takes the sign of the divisor.
This second definition is implemented in some other languages,
such as Python.  Still others, such as Haskell, provide separate
functions for these two definitions.

This patch provides two functions, vnl_math_remainder_truncated
and vnl_math_remainder_floored, for these two definitions,
respectively.  Moreover, they are overloaded to provide a
consistent interface for integer and floating point arguments.

Unit tests have been provided.

Change-Id: Ib474d1591e3019ce1c421ab74325846826ec20d4