COMP: Require NO_MODULE with DCMTK find_package.
As reported in ITK-3409, when DCMTK provides the *.cmake
files, the DCMTK_DIR is set to /usr/lib/cmake/dcmtk, and the subsequent call
to find_package(DCMTK) fails, because it can't find the include directories -
i.e. cmake provided FindDCMTK is used and doesn't try the DCMTK*cmake files.

Force NO_MODULE in the DCMTK find_package calls.  This will require a recent
version of DCMTK that provides a DCMTKConfig.cmake file.

Change-Id: I04dee944a9cd017543e7c223bc8bfd06e226a46e