COMP: Respect CMAKE_CXX_STANDARD during config
During the introspection process it is important
that the tests that run to determine compiler
features use the same compilation flags as
the final program.  This is particularly important
when using the new CMAKE_CXX_STANDARD=11 flag for
C++11 conformance.

Only pass the CMAKE_CXX_STANDARD to try/compile if set.

This is related to: https://cmake.org/Bug/view.php?id=15361#c37760

Change-Id: I71d1f87fdd1218f514a96f0cf626bdff1c0f3081