BUG: Fix multiple external modules wrapping
This commits modifies the generation of the castxml.inc file
needed by CastXML when wrapping an ITK module.
When wrapping multiple external modules from the same Cmake project,
only the include path of the last module was added to the castxml.inc
file. This has been fixed by generating one file per module using
the module name as a prefix.

Change-Id: Ie04c853a7947272185f1013fb6112c9c4d78734b