ENH: Add ITK_CUSTOM_LIBRARY_SUFFIX variable.
ITK libraries have a major and minor suffix version in their names to
prevent DLL hell and to allow multiple versions of ITK to be installed on the
same system.

When using CMake as a build system, the paths are automatically handled by CMake. To assist
building against ITK when not using CMake (not recommended), an
ITK_CUSTOM_LIBRARY_SUFFIX is added to specify the library suffix. If a
suffix is not desired, set this to the empty string.

Change-Id: I6f902144b927f0b379f35e27634487bf7bceeca8