BUG: Fix wrapping build with DEFAULT_MODULES OFF.
Fix the wrapping build when ITK_BUILD_DEFAULT_MODULES is OFF and the only
group that is ON is ITKGroup_Core.

More include's have to be specified explicitly in the wrapping, but this is
required. The check for ITK_BUILD_DEFAULT_MODULES or ITKGroup_Filter in
WrapITKTypes.cmake was not valid because some of these modules are
enabled/required even when these are not explicitly enabled due to module
dependencies.

Change-Id: I35071ff77c80c694de907f3f597edf3fd085a8fb