COMP: Fix CASTXML_EXECUTABLE value for multiple external module
In previous commits a766d34 and 910837d, we were looking for an
already existing CastXML target before adding a new one, in order
to avoid target duplication in the same project.
This commit fixes the value of CASTXML_EXECUTABLE that was set
by every external module even if the target wasn't created.
As a result, we now download, configure and build CastXML in the
binary directory of the first external module of the project.
If more external modules are wrapped, the CASTXML_EXECUTABLE value
will still be pointing at the first binary directory so we can
re-use this CastXML target.

Change-Id: I479066b74ad1fb530e44b2045639e5418070d7c9