PERF:  Avoid temporary std:vector copying.
Previously the 'distances' vector was passed to the
function, and only modified when a temporary vector
was copied into it.  We can avoid the temporary vector,
the memory, and the copy operation by simply using the
passed in vector.

The 'distances' is the storage container used internally for the
NearestNeighbors class.  The 'distances' vector is modified
by the NearestNeighbors class.  By passing in
the 'distances' vector here, we can avoid unnecessary memory
duplications and copy operations.

Change-Id: I9a3b7e71eb128e76cac98e2ff94a5f20ca507db5