ENH: Refactored itkBioCellTest
This patch set refactors the itkBioCellTest.cxx so that the code is more
compact, and more functions of the itk::bio::Cell class get called. This
should improve its code coverage.

The class header includes a CheckPointApoptosis method so that the cell
cycle can be simulated. For now, it just calls the base class same name
function.

Change-Id: I42d8dc6d327612b179f2711fb8ed130d3d3c0a10