COMP: Fix external module testing
This commit follows up on df50151 in which compilation issues related
to itk_add_test were fixed.
Here we remove the ITK_USE_KWSTYLE option from the ITKModuleExternal
module. The logic to set ITK_USE_KWSTYLE is already implemented in the
ITKModuleKWStyleTest module.
It also fixes the BuildKWStyle module to ensure that KWStyle is built
when it has been requested by an external module.

Change-Id: Ieea913c50de1be088b3f67c6145fc7b96f906a8c