COMP: SizeValueType can has different definition for itkArray
SizeValueType for an array is different than SizeValueType
defined in itkIntTypes.h.  We must explicitly define
the type that we want.

Additionally, due to ambiguity introduced by the default
boolean option, if the type of variable pass to the function
does not match SizeValueType exactly, it is unclear if that
type should be cast to bool (matching the first signature)
or to SizeValueType (matching the second signature).

Add explicit versions to allow for disabmiguating the
desired behavior.

Change-Id: I2a29634462aa0de6d3df426f7efc2ca48dcc944d