ENH: added inlined Zero and one values for integer NumericTraits
C++x03 allows inline initialization of static const integers, but also
requires a declaration to resolve linkage. This is now done for
NumericTraits of intrinsic integer types.

For C++11 the values of Zero and One are constexpr and don't require
declaration in a compilation unit.

These values are now available at compilation time for algorithms and
should allow for additional compiler optimization.

Change-Id: Idf4549f29af71685b81b09ebe284fa038f83aae8