COMP: Disabmiguate function calls to SetData
The patch e3cc561c8909d59bb6e8a99b79e2253bcc5a13e1 did
attempted to provide a backwards compatibile solution,
but it was unable to resolve ambiguities in 32/64 bit
with both internal vxl and external old vxl builds.
NOTE: Adding explicit versions of the 3 parameter version
*did not* to allow for disabmiguating the desired behavior.

SizeValueType for an array is different than SizeValueType
defined in itkIntTypes.h.  This causes ambiguities when
a call to SetData(...) is made where the second parameter
is not the same as itk::Array::SizeValueType.

Most naturally, we would expect any integer type to
be automatically typecast to itk::Array::SizeValueType,
but one of signatures for SetData takes 2 arguments
(with the second argument defaulting to false) that implicitly assumes
that the size of the input array is exactly the same
size as the array being passed in.  In this optimization case,
the second argument is bool indicating if the memory should
be managed by the array.  This optimized case conflicts
with the 3 parameter version of SetData (with the third arugment
defaulting to false) when the second argument does not exactly
match itk::Array::SizeValueType.

Additionally, due to ambiguity introduced by the default
boolean option, if the type of variable pass to the function
does not match SizeValueType exactly, it is unclear if that
type should be cast to bool (matching the first signature)
or to SizeValueType (matching the second signature).

SetData was not called anywhere in [Slicer,BRAINSTools,UKF,
ANTs, SlicerExectionModel, AnalyzeObjectMapIO,
HigherOrderAccurateGradient, LesionSizingToolkit, MGHIO,
SCIFIO, SmoothingRecursiveYvvGaussianFilter, SplitComponents]

SimpleITK called SetData in 2 places, and in both cases it
used the fully qualified 3 argument version.

Internal to ITK, there was only 1 place where the
two argument version was used.  Since the two argument version
is such a specialized case with only a single use found in
all source code, it's name was changed to be more specialized
and avoid the ambiguitity.

Change-Id: I0d726efcffa8499556492ccbfae9408cba1d09fe