ENH: Improve itkFEMLoadPoint coverage.
This patch set tries to improve the itk::fem::LoadPoint class' code
coverage. The test is an existing one, but was added to the CMakeLists.txt
file so that it can be called.

The test exercises many more classes, so at some point we will need to
decide whether we should create other individual tests for the classes
involved. BTW, I am not an expert on FEM.

Furthermore, the test is virtually duplicated: two files exist with almost
the same name. The only difference is the use of raw pointers for some
method arguments, and a wrong driver function (in
itkFEMLoadPointUsertest.cxx). If there are no good reasons to keep this
style or the file as a whole, let me know, and I'll delete it in future
patch sets.

Change-Id: I4fea78198b2b363a86d81a82e91d8308ed3c5049