BUG: Install VXL config files to backwards-compatible locations.
This fixes building against the ITK hello world installation example. These
files are not installed into a consistent location with respect to the
upstream VXL location and since the VCL files are just available for compatibility,
modify our source tree to place the config files in the same location prior
to the recent VNL update.

Change-Id: Ic91abae00693a5d04a459b61d416025abbabc6ab