ENH: Provided static code API for external applications
When an external application builds against ITK, the code options used
should be identical as when ITK was built.  Conditional code paths (#if/#endif)
code should be conditioned on values in the itkConfigure.h file, and *NOT*
on compiler internal variables (__cplusplus). This patch identifies
compiler information at configuration time, and then stores that information
for reporting to external users of ITK.

Change-Id: I4114b547b55e3dfb69dd7fd45f7ad1cd28072288