COMP: Build against system HDF5 1.8.16.
In version 1.8.16, HDF5 changed their CMake configuration.

Libraries "hdf5" and "hdf5_cpp" now have the suffix "-shared" or "-static". Add the COMPONENT "shared" or "static" to find_package(HDF5 is now required.

HDF5 now providies an
hdf5-config.cmake which is not compatible with the CMake FindHDF5.cmake.
In this patch "find_package(HDF5" is called with the NO_MODULE option and the
shared or static component specification based on BUILD_SHARED_LIBS.  If the
package is not found, find_package is called again without NO_MODULE so
older versions of HDF5 without a package config file can be found.

In 1.8.16, HDF5 also change the variable name of their CMake include
directories from HDF5_INCLUDE_DIRS to HDF5_INCLUDE_DIR, so both are used for
increased compatibility.

Change-Id: I6fbb2b011eca9a79d40ff711a3cdd3302e7e7de3