BUG: Use LargestPossible region for BSpline domain
Corrected from using the RequestedRegion of the input image to
LargestPossibleRegion for the domain of the BSplineTransform. The
internal pipeline ivar should not be used to define the domain of the
resulting BSpline transform. If the input image was previously used as
input into an Extract image filter then the side-effect of the
RequestedRegion modification causes a change in the behavior of the
BSplineTransformInitializer. This is an unintended side-effect, and
does not follow the standard usage of the pipeline
variables. This has caused problems in the SimpleITK environment.
Additionally, the previous deprecated initializer,
BSplineDeformableTransformInitializer, used the LargestPossibleRegion.

Change-Id: I23e29b852c1dc19fe69f9c79e07f478812daec8f