BUG: Valgrind detected an invalid read
The Math::Compare... methods seem to be accessing too much memory
for some types. In particular RGBPixel<unsigned char>. The tests were
using it.Value() methods to get the values of pixels. This method
returns a reference. Changing the Value() to Get() removes the defect.

Valgrind detected the improper access.

The reverted patch did not fix the underlying issue.

Revert "BUG: Fixed the test itkHDF5ImageIOTest"

This reverts commit c512b28278cb1ef0be1419b1a0ec2eec1e3b802f.

Change-Id: If6fb7911d0220989e5a80e39279f1882cab136e7