COMP: Address warning about this usage in initializer list
Addresses compilation warnings like the following:

itkBSplineDeformableTransform.hxx(34): warning C4355: 'this' : used in
base member initializer list

In the initializer list, "this" has not been initialized and can not
be used.

It is best practice to use static function, in initializer list if
possible. Otherwise, it is recommended style to prefix a member call
with "Self::" to clarify that the method is not being called in a
virtual fashion.

 LocalWords:  virtirual

Change-Id: I5b0e88f69fcc31614834d6ee394d654b2cfddd57