COMP: add ITKCommon_EXPORT to fix link issues with external applications
commit 3ed8e8af5215dd545f0955542cfcd204a9a5218d  in itk NumericTraits leads to compilation error in orfeo toolbox.

I was able to solve them adding ITKCommon_EXPORT macro in the declaration. Not sure to understand why it is necessary as the related commit in ITK seems to handle linkage resolution.

The link error related to new declaration of inlined Zero for integers type can be seen on otb dashboard:

http://dash.orfeo-toolbox.org/viewBuildError.php?buildid=225288

Configurations for ITK and OTB which leads to the linking issue can also be found on otb git repositories:

itk configuration:

https://git.orfeo-toolbox.org/otb-devutils.git/blob/HEAD:/Config/pc-christophe/pc-christophe-ITK-trunk.cmake

otb configuration:

https://git.orfeo-toolbox.org/otb-devutils.git/blob/HEAD:/Config/pc-christophe/pc-christophe-OTB-Nightly-clang-ThirdPartyTrunk.cmake

Change-Id: I3b528d2ccdf61c061b84f3ed5417a6eeb771f3dc