BUG: Increase MaskLabel backwards compatibility in N4 bias correction.
This is a follow-up to 91624a11d01cddbb4a986cf7b2c29526744971ec where the
MaskLabel ivar was removed from N4BiasFieldCorrectionImageFilter. The previous
API and behavior is restored. An additional ivar UseMaskLabel, which
defaults to 'true' for backwards compatibility, is added.  If this is false,
then any non-zero voxel in the mask image is considered part of the mask, and
the MaskLabel is not used.

Change-Id: I0ef9067a7297b9829ca902ac4c48ee7b3f9263aa