COMP: Hack HDF5 to build under ASan & UBSan
Parts of HDF are built as part of ITK’s build process.
It crashes and burns under ASan & UBSan thus preventing all
of ITK from being built & tested with these invaluable tools.
Filed as HDF bug HDFFV-8147 in 2012, but the HDF folks don't
seem willing or able to fix it.

Change-Id: I84dda62e7bdcbbd5f31e517491726df64c34e723