ENH: Test address of NumericTraits One and Zero constexpr
Fix the test to check the One and Zero values.

Added check that the address of One and Zero, can be make. This
ensures that the variable is declared and has an address. Note, that
we are not checking that this address is universal and unique, just
that is exists.

Change-Id: Ia9541c8c08fa30bccb1d6b53fe3d0ea8814e5ad9