BUG: Wrapping intermediate files were not automatically updated
When wrapping ITK, CastXML converts the source code into an XML abstract syntax.
This XML needs to be updated if the source code is modified. CastXML is run
as part of an 'add_custom_command' command. This command did not depend on the header
file that is being wrapped.
One solution to depend on the header you want to wrap, is to add the 'IMPLICIT_DEPENDS'
argument in the 'add_custom_command' [1]. Unfortunately this argument is currently only supported for
Makefile generators, so a more general solution needed to be found.
All the source and header files that are part of the filter that is wrapped are being added as
a dependency of the custom command. If one of them is modified, the custom command will
be run again.
[1] https://cmake.org/cmake/help/v3.5/command/add_custom_command.html

Change-Id: I5bee60574266a0cf22d4fce2d38ef66c36ad994a