ENH: Enable run-time dependency on Python library
Add a try compile to determine if the linker supports unresolved
symbols, such as on Linux with ELF binaries. If so don't link to
libpython, but use the runtime environment of the python executable to
resolve the symbols. Continue usage of -undefined dynamic_lookup on
OSX to enable simular behavior.

Change-Id: I598835a173560fb90c94738efd4d7f5618c64914