BUG: Interpretation of the Euler angles ZYX or ZXY was not exported.
The flag indicating whether the Euler angles should be composed in ZYX
or ZXY order was not part of the fixed parameters as it should have
been. This patch adds it to the fixed parameters. This was an issue
with serialization de-serialization of the transform. The choice of
composition method was not serialized and writing followed by reading
changed the transform.

Change-Id: I0a2230022c35b8a7ef2c5c4d91e7aabfa50b9d7c