BUG: demonstrate dynamic_cast failure when reading transforms
This test fails with some compilers, such as OSX clang 7.3.0 with
shared libraries. The dynamic_cast fails because there are multiple
instantiations of the Transform objects in the runtime with different
typeinfo and vtables. Because some of these symbols are private, some
compilers don't recognize them as the same object.

Change-Id: I63affda401fcf2d9641e095ecec02b40b30061ff