BUG: Fix Python wrapping on Windows when ITK_USE_64BIT_IDS is ON.
New wrapping variables are introduced, e.g. ITKM_ULL ITKM_SLL for unsigned
long long and signed long long. Also ITKM_IT, ITKM_ST, ITKM_OT. These are to
define variables to use when itk::IdentifierType, itk::SizeValueType, and
itk::OffsetType. Mimicking the logic in itkIntTypes.h, these resolve to the
long long versions when on Windows and when ITK_USE_64BIT_IDS is ON.

These variables are used throughout the wrapping to address the Windows build
with ITK_USE_64BIT_IDS.

The wrapping always assumes that the toolchain used for wrapping supports
'unsigned long long' and 'signed long long'. This will have to be revisited in
the future if it is an issue.

Change-Id: I8b3dcf22d2fd1d0cd0e071e7916d1a0c7fc1a4f5