ENH: Adding GetArrayFromImage in python itk namespace
The function GetArrayFromImage is part of PyBuffer.
When calling PyBuffer, one has to specify the type of
the input image they work with. GetArrayFromImage that
is directly in itk namespace finds the type of the
image that is given and instantiates a PyBuffer object
of the matching type to convert the ITK image into
a numpy array.

Change-Id: Ieac3a7934d58861d6dfa2ade846c8d90ffdfb99e