ENH: Adding GetImageFromArray in python itk namespace
The function GetImageFromArray is part of PyBuffer.
When calling PyBuffer, one has to specify the type of
the output image they expect. GetImageFromArray that
is directly in itk namespace finds the type of the
elements of the array that is given and instantiates
a PyBuffer object of the matching type to convert the
numpy array into an ITK image.

Change-Id: I2688dc0295dc4dcbf7d54410cdc3acd37fd883b7