ENH: *_WRAP_* variable per external module
Instead of using global ITK_WRAP_PYTHON, ITK_WRAP_JAVA,
ITK_WRAP_* variables for all projects, variables per project
are created. These variables depend on the corresponding
global ITK_WRAP_* variable, meaning that if ITK was not
wrapped in one language, it is not possible to wrap the
external module in that language. A CMake option
is created for each module for each language ITK was wrapped
in. Therefore only meaningful variables are exposed to the
user. Module specific variables are initialized with the
corresponding global variable value.

Change-Id: Ib1bd64be90c14aef0e95d011b85cd24196f19520