COMP: Devirtualize private methods to prevent instantiation
This patch addresses the following compilation warning with GCC 6.1:

itkMinMaxCurvatureFlowFunction.hxx:336:43: warning: array subscript is
above array bounds

This class have as an overloaded method used for different image
dimensions. The correct one will be called, but the un-used methods
would cause invalid access to arrays. Method marked as virtual are
instantiated and placed in the vtable for an object, event if they are
not used. Since these methods are private they can only be used by
it's self. We safely remove the virtual qualifier to only instantiate
methods needed.

Change-Id: Ifa2c06376c3913debe93801194aff65e035b91c9