BUG: Fix OpenCV-2.4.9 camera empty image issue.
The itkOpenCVVideoIOTest.cxx fails for a particular local setting when
using the OpenCV 2.4.9 version, and when the local camera is addressed
with DirectShow drivers.

The test specifically fails when querying the number of channels of the
tempImage variable, which is empty.

This patch set modifies the test so that an exception is thrown if
tempImage is is empty.

This patch set removes the unnecessary method documentation from the
itk::OpenCVVideoIO class implementation file.

Change-Id: I45e6421d13daa0ebd15d318f3e6094e0f21e095d