BUG: Decrement index in BSplineTransform by ULPs.
As reported by Denis Shamonin in ITK-3425:

1. How crash was found.
We have found the rare but reproducible crash during the resampling in combination with BSplineTransform.
Due to the difficult nature of the crash, we are providing the full minimum test setup and the data to reproduce it. (See attachment BSplineTransformTest archive)
To reproduce our findings, please compile and run the tests from archive (ctest -VV).
Crash happens during the following call in itkResampleImageFilter.hxx call inputPoint = transformPtr->TransformPoint(outputPoint).
We studied the crash in details. The crash was introduces during the following commit:
SHA-1: 9ea5521e0fd2ed30404a72c53b1bdb15ab51a057
COMP: Removes floating-point equality checks
Equality checks between floating point values and other types are
sometimes required to be exact and other times need only be approximate.
Here direct equality checks are replaced with functions calls to either
an exact or approximate comparison. Using these functions explicitly
stating which kind of comparison is being requested improves clarity and
readability of code, and removes compiler warnings ie. Wfloat-equal.

2. The nature of the crash.
The crash caused by incorrectly computing the inside valid region during the BSplineTransform transform point call.
BSplineTransform::InsideValidRegion(). The following call will compares two float numbers to be equal.
ITK git
if(itk::Math::AlmostEquals(index[j], maxLimit))
{
  index[j] -= 1e-6;
}
In our test the index[0]=5.00000143 and maxLimit=5.00000000. the itk::Math::AlmostEquals will return true.
In fact this two float numbers are 3 ULPs (units in the last place) apart, and itk::Math::AlmostEquals will check if the two floating point numbers are less than (or equal to) 4 ULP apart.
Then the constant number 1e-6 will be subtracted resulting in index[0]= 5.00000048.
This is not enough, it should go to 4.99999... !
This not correct subtraction will trigger a crash later during construction of the itk::ImageScanlineConstIterator. Because an index outside the image will be requested, which does not exist.
The core of the issue is that in some cases 1e-6 < 4ULP
So, in some cases subtracting the constant 1e-6 is enough and a larger number >4ULP should be subtracted.

Change-Id: Iaa0eea59827a45578d4e9b935916109de2ff4a09