STYLE: Improve GACLevelSet classes' style
This patch set improves the GeodesicActiveContourLevelSet segmentation
classes' style:
Remove unnecessary/duplicated method documentation in various
implementation files.
Move the PrintSelf method implementation of the
itk::GeodesicActiveContourLevelSetFunction class from the header file to
the implementation file.
Move the PrintSelf method implementation of the
itk::GeodesicActiveContourShapePriorLevelSetFunction class from the header
file to the implementation file.
Change the method body comment style to //.
Move the PrintSelf method in
itk::GeodesicActiveContourShapePriorLevelSetImageFilter to the end of the
implementation file.
Improve the comment in the protected copy constructor and assignment
operator overload in
itk::GeodesicActiveContourShapePriorLevelSetImageFilter class.

Change-Id: Id70cd98b059ff0693afb24fcca18bf8e57019dfe